Updating tests, and fixing namespace bug in workflow
